requirements.txt
requests
fastapi
uvicorn
google-auth
faker
psycopg2-binary